Allow CH4 pseudoflux to be calculated and archived in GCHP
Calculation of the CH4 pseudoflux was implemented by GMAO, but
walled off by C-preprocessor statements. Calculating and using
this flux is a useful feature for GCHP and there is no reason
to not have it available, so these statements have been removed.
The CH4 pseudoflux can now be calculated for any full-chemistry
simulation as long as the CH4 emissions flag is set to "false"
and the CH4pseudoflux diagnostic is requested through HISTORY.rc.

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>