Limited the effect of the CH4 emissions flag
With CH4 pseudofluxes now calculable, users need a way to then specify
those pseudofluxes as an emission. This edit means that this can now
be performed by setting the "CH4 emissions" flag in input.geos to
True and then reading in the pseudofluxes like any other emission,
through HEMCO. A redundant output statement was also removed as it
referred to a flag which is no longer used or even settable.

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>