Fix typo to re-enable species not in restart
A CYCLE command which should only be used in GEOS was instead
being used for both GCHP and GEOS, due to an error during an
earlier merge. This has been fixed. An incorrect am_I_Root
usage (where am_I_Root is not yet set) has also been replaced
with a call to MAPL_am_I_Root in the same file.