Use MAPL_UNDEF to fix RRTMG output averaging
RRTMG is called only infrequently. To ensure that the averaging
is performed correctly, we can use MAPL_UNDEF - setting a diag
field to MAPL_UNDEF will mean that it isn't used when calculating
averages. With this fix, using time-averaged fields for RRTMG
(with the same output frequency as the RRTMG call frequency) will
give the correct output. Longer averaging periods will also work.