Enable read-in of dynamical heating for RRTMG
This is the first fully functional commit of the FDH assumption.
GCHP can now be configured to either write out the dyn heating
(Read dynamical heating = F in input.geos and IMPORT_DYN_HEATING
set to 0 in GCHP.rc) or to read it in. In the former case, the
entry for DynHeating in ExtData.rc is ignored; in the latter
case, it must be present. This commit also brings some small
quality of life improvements.