Update run directories for RRTMG SEFDH
The standard run directories are now updated to include the two
new lines for RRTMG (FDH and SEFDH). The GCHP HISTORY.rc template
also now includes the collection necessary for outputting
dynamical heating from RRTMG (needed for either FDH or SEFDH),
and has the new fields for tropopause RF and for CO2, CFC, H2O,
and N2O forcing to be calculated. ExtData.rc now includes the
DynHeating field, which is only needed if IMPORT_DYN_HEATING is
set to 1 in GCHP.rc (set by default to 0 in the updated GCHP.rc).