Allowing simulations with low tops or low level counts
The XLAIMULTI approach (pretending that the 73 LAI arrays are vertically stacked),
while efficient, causes an error to be thrown if the number of model levels is
less than 73. Until a better solution can be found, this has been reverted to
reading in LAI on each level separately. Another change was needed in the
linoz code since it assumes that the model top will be above 0.3 hPa.